+++ benches/storage_sparse.rs
@@ -113,13 +113,13 @@ macro_rules! tests {
 
             impl Default for CompInt {
                 fn default() -> Self {
-                    CompInt(0)
+                    Self(0)
                 }
             }
 
             impl Default for CompBool {
                 fn default() -> Self {
-                    CompBool(true)
+                    Self(true)
                 }
             }