+++ examples/saveload.rs
@@ -9,7 +9,8 @@ use specs::{
     error::NoError,
     prelude::*,
     saveload::{
-        DeserializeComponents, MarkedBuilder, SerializeComponents, SimpleMarker, SimpleMarkerAllocator,
+        DeserializeComponents, MarkedBuilder, SerializeComponents, SimpleMarker,
+        SimpleMarkerAllocator,
     },
 };
 
@@ -114,7 +115,7 @@ fn main() {
     // This predifined marker uses a `HashMap<u64, Entity>` to keep track of all
     // entities that should be (de)serializable, as well as which ids are
     // already in use.
-    world.add_resource(SimpleMarkerAllocator::<NetworkSync>::new());
+    world.insert(SimpleMarkerAllocator::<NetworkSync>::new());
 
     world
         .create_entity()