+++ src/changeset.rs
@@ -1,13 +1,8 @@
 //! Provides a changeset that can be collected from an iterator.
 
-use hibitset::BitSet;
 use std::{iter::FromIterator, ops::AddAssign};
 
-use crate::{
-    join::Join,
-    storage::{DenseVecStorage, UnprotectedStorage},
-    world::{Entity, Index},
-};
+use crate::{prelude::*, storage::UnprotectedStorage, world::Index};
 
 /// Change set that can be collected from an iterator, and joined on for easy
 /// application to components.
@@ -40,11 +35,18 @@ use crate::{
 /// }
 /// # }
 /// ```
-#[derive(Derivative)]
-#[derivative(Default(bound = ""))]
 pub struct ChangeSet<T> {
     mask: BitSet,
     inner: DenseVecStorage<T>,
+}
+
+impl<T> Default for ChangeSet<T> {
+    fn default() -> Self {
+        Self {
+            mask: Default::default(),
+            inner: Default::default(),
+        }
+    }
 }
 
 impl<T> ChangeSet<T> {
@@ -90,7 +92,7 @@ where
     T: AddAssign,
 {
     fn from_iter<I: IntoIterator<Item = (Entity, T)>>(iter: I) -> Self {
-        let mut changeset = ChangeSet::new();
+        let mut changeset = Self::new();
         for (entity, d) in iter {
             changeset.add(entity, d);
         }