+++ src/prelude.rs
@@ -7,8 +7,8 @@ pub use crate::join::Join;
 pub use crate::join::ParJoin;
 pub use hibitset::BitSet;
 pub use shred::{
-    Accessor, Dispatcher, DispatcherBuilder, Read, ReadExpect, RunNow, StaticAccessor, System,
-    SystemData, World, Write, WriteExpect,
+    Accessor, Dispatcher, DispatcherBuilder, Read, ReadExpect, Resource, ResourceId, RunNow,
+    StaticAccessor, System, SystemData, World, Write, WriteExpect,
 };
 pub use shrev::ReaderId;
 
@@ -20,8 +20,8 @@ pub use shred::AsyncDispatcher;
 pub use crate::{
     changeset::ChangeSet,
     storage::{
-        ComponentEvent, DenseVecStorage, FlaggedStorage, HashMapStorage, NullStorage, ReadStorage,
-        Storage, Tracked, VecStorage, WriteStorage,
+        ComponentEvent, DefaultVecStorage, DenseVecStorage, FlaggedStorage, HashMapStorage,
+        NullStorage, ReadStorage, Storage, Tracked, VecStorage, WriteStorage,
     },
     world::{Builder, Component, Entities, Entity, EntityBuilder, LazyUpdate, WorldExt},
 };