+++ src/world/tests.rs
@@ -96,7 +96,7 @@ fn lazy_execution() {
 #[test]
 fn lazy_execution_order() {
     let mut world = World::new();
-    world.add_resource(Vec::<u32>::new());
+    world.insert(Vec::<u32>::new());
     {
         let lazy = world.read_resource::<LazyUpdate>();
         lazy.exec(move |world| {